requirements.txt

langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.40
langchain-google-genai==2.0.5
langchain-pinecone==0.2.3
langchain-text-splitters==0.3.2


fastapi==0.110.0
uvicorn==0.29.0



redis==5.0.3  


pyjwt==2.8.0  


python-dotenv==1.0.1

requests==2.31.0
